Croak if a user tries to specify an invalid protocol.
The previous revision of the code would fallback to udp if the user's
choice of protocol was incorrect.  This patch checks the validity of the
protocol and calls croak() when a bad protocol string is sent to the
constructor.  Added test to ensure the module croaks() when there's a
bad protocol specified. Named some unnamed tests.