Enable support for multiple statsd endpoints. (#2)
* Enable support for multiple statsd endpoints.

In order to ensure long downtimes for statsd endpoints, or to aggregate
statsd metrics at multiple levels, it is necessary to send data to
multiple endpoints.  This is accomplished while maintaining backwards
compatbility.

    my $statsd = Etsy::StatsD->new();
    my $statsd = Etsy::StatsD->new( 'statagg1' );
    my $statsd = Etsy::StatsD->new( 'statagg1', 8130 );

Continue to work.  New behavior is added:

    my $statsd = Etsy::StatsD->new(['localhost', 'statagg1']);
    # Sends all metrics (with the same results from sampling) to
    # localhost:8125 and the statagg1:8125.

    my $statsd = Etsy::StatsD->new(['localhost:8125', 'statagg1', 'statagg2' ]);
    # To: localhost:8125, statagg1:8130, statagg2:8130

There are two "real world" use cases:

    Host local statsd instance to aggregate in a host metrics namespace:
        statsd.hosts.<hostsname>.counters ..

    Datacenter specific aggregator:
        statsd.dc.<datacenter>.counters ..

    Global aggregator:
        statsd.global.counters ..

Because the replication is being done at this level, not by wrapping the
Etsy::StatsD objects, the sampling rates are consistent across the
views.  It's still UDP, so there will be loss, but we need to start with
a level playing field to ensure all levels of aggregation are playing
with the same data sets.

Another use is for HA on statsd.  Reboots for kernel upgrades are common
practice and having the ability to down an aggregator is key.  You can
now specify multiple hosts and those aggregators will be sent the same
data. Since graphite operates on a "last metric for bucket wins", we can
send duplicate data without corrupting the system, ie:

    Client -> StatsD Aggregater ---> Graphite
         `--> StatsD Aggregater _/

There's some hand waiving around which statsd instance will be the last
and misalignment with the "buckets" but this is as close to the sun we
can fly without getting burnt with high levels of complexity.

* Incorporating feedback on the pull-request.

Ensure older perls are supported by migrating to ||= instead of //=.
Remove the call to keys() in list context to avoid the overhead of
copying a large array of potential keys from an update.  Call keys() in
void context to reset the each() iterator on the %$sampled_data hash.

* Satisfy all possible world views on how to ship metrics to statsd.

This patch adds a third, optional, argument to the address specification
allowing TCP end points for the Etsy::StatsD clients.  To enable TCP,
the user must use the form "<host>:<port>:<proto>" in the host parameter
position.  This is a minor inconvenience to enable all the things!

* Croak if a user tries to specify an invalid protocol.

The previous revision of the code would fallback to udp if the user's
choice of protocol was incorrect.  This patch checks the validity of the
protocol and calls croak() when a bad protocol string is sent to the
constructor.  Added test to ensure the module croaks() when there's a
bad protocol specified. Named some unnamed tests.

* croak() instead of carp() when there's a bad protocol.

There are two checks, one for an array reference and one for the host as
a string.  The array ref had the croak() call correct, but the string
check was only carp()ing.  Add a test to catch this failure too.

* Drop support for <host>:<port>:<proto> when the HOST parameter is a
scalar.  If users require the new functionality, then they must pass an
array reference in to get the advanced parsing.